+++ tests/test_audiotools.py
@@ -408,6 +408,7 @@ class test_oaudio(unittest.TestCase):
         l_tone = 20*np.log10(np.sqrt(0.5) / 20e-6)
         assert audio.calc_dbspl(sig) == l_tone
 
+
     def test_set_dbsl(self):
         fs = 100e3
         signal = audio.generate_tone(100, 1, fs)
@@ -437,6 +438,10 @@ class test_oaudio(unittest.TestCase):
         signal = audio.set_dbfs(signal, -5)
         assert(signal.max() == m)
 
+        assert audio.set_dbfs(2, 0, norm='peak') == 1
+        signal = audio.generate_tone(1000, 8, 48000)
+        assert audio.set_dbfs(signal, 0, 'peak').max() == 1
+        assert audio.set_dbfs(signal, -3, 'peak').max() == 10**(-3/20)
 
     def test_phon_to_dbspl(self):
         # Test some specific Values
@@ -694,19 +699,6 @@ class test_oaudio(unittest.TestCase):
         signal[signal < 0] = 0
         c = audio.crest_factor(signal)
         testing.assert_almost_equal(c, 20*np.log10(2))
-
-
-    # def test_phaseshift(self):
-    #     signal = audio.generate_tone(100, 1, 100e3)
-    #     signal2 = audio.generate_tone(100, 1, 100e3, np.pi)
-    #     signal3 = audio.phase_shift(signal, np.pi, 100e3)
-    #     testing.assert_almost_equal(signal2, signal3)
-
-    #     signal1 = audio.generate_tone(100, 1, 100e3)
-    #     signal2 = audio.generate_tone(200, 1, 100e3, np.pi)
-    #     signal = np.column_stack([signal1, signal2])
-    #     signal = audio.phase_shift(signal, np.pi, 100e3)
-
 
     def test_band2rms(self):