+++ CHANGELOG.md
@@ -2,6 +2,9 @@ Develop branch
  - Moved filterbanks from audiotools.filter to audiotools.filter.bank submodule
  - Added a default gammatone filterbank: filter.auditory_gamma_bank
  - audiotools.freqarange now supports octave spacing
+
+0.62 -> 0.64
+ - Added parameter to auditools.dbfs to specify if dB fullscale is relative to peak or rms level
  - Deleted the long depricated Signal.add_noise_noise method
  - Implemented Signal.stats submodule
  - Added deprecationwarning to Signal.calc_dbfs and Signal.calc_dbspl which
@@ -10,6 +13,7 @@ Develop branch
  - Added the audiotools.cmplx_corr function which calculates the
    complex-valued correlation coefficent
  - Renamed calc_coherence to cmplx_crosscorr
+ - Fixed a bug in signal.phase_shift that sometimes resulted in a complex valued signal
 
 0.61 -> 0.62
  - Fixed the shape of the channels after appling a filterbank